Still needs a lot of changes, but commited changes so far for the home page.
