+++ package.json
@@ -28,7 +28,6 @@
     "next-themes": "^0.4.4",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
-    "react-icons": "^5.5.0",
     "swiper": "^11.2.6",
     "swr": "^2.3.2",
     "tailwind-merge": "^3.0.1",
