+++ src/components/home/FeaturedNetworksButton.tsx
@@ -6,8 +6,6 @@ import {
   DropdownMenuContent,
   DropdownMenuItem,
   DropdownMenuTrigger,
-  DropdownMenuLabel,
-  DropdownMenuSeparator,
 } from '@/components/ui/dropdown-menu'
 import { ChevronDown, Loader2, ExternalLink } from 'lucide-react'
 import { useFeaturedNetworks } from '@/hooks/use-content-service'
