+++ src/components/search/SearchBox.tsx
@@ -1,10 +1,13 @@
 'use client'
 
+import { Input } from '@/components/ui/input'
+import { Button } from '@/components/ui/button'
+import { SearchIcon } from 'lucide-react'
+
 import { useEffect, useRef, useState } from 'react'
 import { useRouter, useSearchParams } from 'next/navigation'
 import { useSearchStore } from '@/stores/search-store'
-import { Search } from 'lucide-react'
-import { Input } from '@/components/ui/input'
+import { FeaturedNetworksButton } from '@/components/search/FeaturedNetworksButton'
 
 export function SearchBox() {
   const { setQuery } = useSearchStore()
@@ -14,27 +17,29 @@ export function SearchBox() {
 
   const inputRef = useRef<HTMLInputElement>(null)
 
-  // If there's a query in the URL, sync it with our state on mount
   useEffect(() => {
     const urlQuery = searchParams.get('q')
     if (urlQuery) {
+      // Decode the URL query and set it to the input
       setCurrentQuery(decodeURIComponent(urlQuery))
     }
   }, [searchParams])
 
-  // Handle Enter key or submit button
   const handleSubmit = (e: React.FormEvent) => {
     e.preventDefault()
+
     const nextQuery = currentQuery.trim()
+    // Early return if query is empty - this prevents any submission logic
     if (!nextQuery) {
       console.log('Empty search prevented')
+
       return
     }
+
     setQuery(nextQuery)
     router.push(`/search?q=${encodeURIComponent(nextQuery)}`)
   }
 
-  // Prevent blank "Enter"
   const handleKeyDown = (e: React.KeyboardEvent<HTMLInputElement>) => {
     if (e.key === 'Enter' && !currentQuery.trim()) {
       e.preventDefault()
@@ -42,57 +47,52 @@ export function SearchBox() {
     }
   }
 
-  // Clear search if the input is cleared via the built-in browser 'x' in type=search
   useEffect(() => {
     const input = inputRef.current
     if (!input) return
 
     const handleSearch = () => {
+      // This will only fire when the search is cleared via the "x" button
       if (input.value === '') {
         console.log('Search cleared via x button')
         setCurrentQuery('')
+        // Optional: Also clear the store and URL
         setQuery('')
         router.push('/search')
       }
     }
+
     input.addEventListener('search', handleSearch)
     return () => input.removeEventListener('search', handleSearch)
   }, [setQuery, router])
 
+  useEffect(() => {}, [currentQuery])
+
   return (
-    <form onSubmit={handleSubmit} className="w-full max-w-7xl">
-      {/* Rounded container for the search input + icon */}
-      <div className="flex items-center rounded-lg border border-gray-300 pl-1 pr-3 py-0.5">
-        {/* The text input */}
+    <form onSubmit={handleSubmit} className="flex items-center space-x-2">
+      <div className="relative border-1 border-slate-500 rounded-sm overflow-hidden">
         <Input
           ref={inputRef}
           type="search"
-          placeholder="Search for networks, users, and groups"
+          placeholder="Search for networks..."
           value={currentQuery}
           onChange={(e) => setCurrentQuery(e.target.value)}
           onKeyDown={handleKeyDown}
-          className="
-            flex-1
-            border-none
-            bg-transparent
-            placeholder-gray-500
-            outline-none 
-            focus:outline-none 
-            focus:ring-0 
-            focus:border-transparent
-          "
+          className="w-80 md:w-96 lg:w-[20rem] [&::-webkit-search-cancel-button]:appearance-auto [&::-webkit-search-cancel-button]:h-5 [&::-webkit-search-cancel-button]:w-5 [&::-webkit-search-cancel-button]:cursor-pointer"
         />
-        {/* Vertical divider */}
-        <div className="mx-2 h-5 w-px bg-gray-300" />
-        {/* Search icon */}
-        <button
+      </div>
+      <div className="flex items-center gap-2">
+        <Button
+          className="border-slate-600"
           type="submit"
-          aria-label="Search"
-          className="text-gray-400 hover:text-gray-600 disabled:hover:text-gray-400 disabled:opacity-50"
+          variant="outline"
           disabled={!currentQuery.trim()}
         >
-          <Search className="h-5 w-5" />
-        </button>
+          <SearchIcon />
+        </Button>
+      </div>
+      <div className="flex items-center gap-2">
+        <FeaturedNetworksButton />
       </div>
     </form>
   )
